check video on CI in allure report

+++ quality-assurance/src/test/java/pl/kargolek/extension/VideoRecordExt.java
@@ -28,6 +28,11 @@ public class VideoRecordExt implements BeforeEachCallback, AfterEachCallback {
     @Override
     public void afterEach(ExtensionContext extensionContext) throws Exception {
         stopDockerContainer("quality-assurance-chrome_video-1");
+
+        if(!tempVideo.exists()){
+            throw new RuntimeException("Unable to locate temp video");
+        }
+
         var renamedVid = new File(System.getProperty("user.dir") + "/target", "video-" + RandomUtils.nextLong() + ".mp4");
         tempVideo.renameTo(renamedVid);
         try {
